now using PNG files instead of image(6)s. bullet TTL taken into account during stepping.

+++ dat.h
@@ -124,7 +124,7 @@ struct Particle
 struct Bullet
 {
 	Particle;
-	ulong ttl; /* in s */
+	double ttl; /* in s */
 	int fired; /* XXX: |v| != 0 */
 };
